requirements.txt
sentence-transformers==3.4.1
langchain-huggingface==0.1.2
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain_openai
langchain_experimental
-e .